+++ src/gs/gsTile.h
@@ -32,8 +32,7 @@ namespace gs
     protected:
         static int                              idCounter;
         Type                                    surface;
-        //const gs::Vec3f                         centroid;
-        const gs::Vec3f                         normal;
+        const gs::Vec3d                         normal;
         const double                            height;
 
         const vector<shared_ptr<gs::Vertex>>    vertices;
